pyproject.toml
[tool.poetry]
name = "gpt-snippet-generator"
version = "0.1.0"
description = ""
authors = ["alejoboga20 <alejoboga19@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "<=1.0.0"
langchain = "^0.0.353"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
